Added all the jump long instructions.
Changed the jump and links so that they are now working properly, by saving the current programcounter in the source, not linking back.

Corrected the jumps so they point at the right program counter, as they are ahead by a program counter.

CarryBit has been added to all subtract, adds and logical shifts.

There is also a operationnumbers.txt which you can filter of the operationnumber in GTKwave so you can read the opcode isntead of the number.

It is now a (hopefully) working 16/32bit processor.